+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `mkDerivation` now uses the default shellCommands in `base.mkShellCommands` if none are supplied.
 - Terraform components can define shellCommands.
 - Generated setup.py files in python components excludes the generated tests folder.
+- documentation derivations now merges `shellCommands` correctly.
+- Python components' `nativeBuildInputs` and `checkInputs` now resolve lists properly
+  again.
 
 ### Changed
 - shellcheck from the ci attribute will now find more shell scripts and is faster.